Fix int64 JSON serialization in Chinese character metrics

+++ src/prime_rl/orchestrator/orchestrator.py
@@ -601,14 +601,14 @@ async def orchestrate(config: OrchestratorConfig):
             },
             # Env metrics
             **{f"metrics/{metric}": metrics_df[metric].mean() for metric in metrics_df.columns},
-            # Chinese character metrics
-            "chinese/char_count": chinese_df.chinese_chars.sum(),
+            # Chinese character metrics (cast to native Python types for JSON serialization)
+            "chinese/char_count": int(chinese_df.chinese_chars.sum()),
             "chinese/char_ratio": (
-                chinese_df.chinese_chars.sum() / chinese_df.total_chars.sum()
+                float(chinese_df.chinese_chars.sum() / chinese_df.total_chars.sum())
                 if chinese_df.total_chars.sum() > 0
                 else 0.0
             ),
-            "chinese/rollout_ratio": chinese_df.has_chinese.mean(),
+            "chinese/rollout_ratio": float(chinese_df.has_chinese.mean()),
             # Time metrics
             "time/step": step_time,
             "time/generate_completions": generate_completions_time,
